Some tweaks

+++ src/api/bottleofench/main/Factorial.java
@@ -7,9 +7,9 @@ public class Factorial {
     public static void main(String[] args) {
         System.out.println("Введите число, факториал которого вы хотите получить:");
         Scanner scanner = new Scanner(System.in);
-        int number = scanner.nextInt();
-        int output = 1;
-        for (int i = 1; i <= number; i++) output *= i;
+        long number = scanner.nextLong();
+        long output = 1;
+        for (long i = 1; i <= number; i++) output *= i;
         System.out.println(output);
     }
 }
